+++ lib/thainum.dart
@@ -41,6 +41,7 @@ export 'src/money.dart';
 export 'src/numerals.dart';
 export 'src/parse.dart';
 export 'src/percent.dart';
+export 'src/short.dart';
 export 'src/speak.dart';
 export 'src/spell.dart' hide isDigits, splitDecimalInternal, DecimalParts;
 export 'src/thai_id.dart';